// src/app/base/models/advanced-search/advanced-search-operator.enum.ts
export enum AdvancedSearchOperator {
    CONTAINS = "CONTAINS",
    ENDS_WITH = "ENDS_WITH",
    EQUALS = "EQUALS",
    IS_GT = "IS_GT",
    IS_GTE = "IS_GTE",
    IS_LIKE = "IS_LIKE",
    IS_LT = "IS_LT",
    IS_LTE = "IS_LTE",
    NOT_EQUALS = "NOT_EQUALS",
    STARTS_WITH = "STARTS_WITH",
    LATEST_DATE = 'LATEST_DATE',
    IS_DATE_GTE = 'IS_DATE_GTE',
    IS_DATE_LTE = 'IS_DATE_LTE',
    IS_DATE_LT = 'IS_DATE_LT',
    IS_NULL = 'IS_NULL',
    RES_ID_AND_ROLE = 'RES_ID_AND_ROLE',
    SKILL_RATING_IS_GTE = 'SKILL_RATING_IS_GTE',
    HOURLY_COST_IS_LTE = 'HOURLY_COST_IS_LTE',
    HOURLY_COST_IS_GTE = 'HOURLY_COST_IS_GTE'
}